refactor: move model code to its own file
